updates for paired-end datasets
added rule for RSEM (left out stranded option for now as test data seems to be unstranded). added rsem path and reference path to config file

added twopassmode and output transcriptomeSAM to STAR to make compatible with RSEM

added rule for cutadapt (STAR failed on the test data due to some reads having a length 0, which need to be removed with cutadapt before mapping, also though it would be good to have some general quality trimming included)

test data for paired end seem to have base qualities of 2, so including -q 20 in cutadapt caused it to fail. may need to change test data to address.

added temporary paths for cutadapt and multiqc from one of my conda directories, which can be changed in future iterations. ultimately may be easiest to provide pipeline with a yml file for a conda env that contains all of the required software tools

updated picard rRNA list in config file as there was a sequence dictionary clash