Update config.yaml
removed the slurm option for specifying an email address. In the manual for sbatch it states that if this is not specified the default is the submitting user, which I think is a better option. 

From the manual

--mail-user=<user>
User to receive email notification of state changes as defined by --mail-type. This may be a full email address or a username. If a username is specified, the value from MailDomain in slurm.conf will be appended to create an email address. The default value is the submitting user.